+++ src/ft_strncat.c
@@ -6,16 +6,16 @@
 /*   By: sungurea <sungurea@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/16 21:47:42 by sungurea          #+#    #+#             */
-/*   Updated: 2018/08/08 23:07:03 by sungurea         ###   ########.fr       */
+/*   Updated: 2018/08/10 17:34:21 by sungurea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 
-char	*ft_strncat(char *dest, char *src, int nb)
+char	*ft_strncat(char *dest, char *src, size_t nb)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	while (dest[i])
